+++ src/OrderLinesList/OrderLinesFiltersContainer.js
@@ -1,4 +1,4 @@
-import React, { useCallback } from 'react';
+import React from 'react';
 import PropTypes from 'prop-types';
 import { get } from 'lodash';
 
@@ -7,28 +7,22 @@ import {
   DICT_FUNDS,
   fundsManifest,
 } from '@folio/stripes-acq-components';
-import OrderLinesFilters from '@folio/plugin-find-po-line/FindPOLine/OrderLinesFilters';
+import { OrderLinesFilters } from '@folio/plugin-find-po-line';
 
 import {
   MATERIAL_TYPES,
 } from '../components/Utils/resources';
 
-const applyFiltersAdapter = (applyFilters) => ({ name, values }) => applyFilters(name, values);
-
 const OrderLinesFiltersContainer = ({ resources, activeFilters, applyFilters, disabled }) => {
   const funds = get(resources, `${DICT_FUNDS}.records`);
   const materialTypes = get(resources, 'materialTypes.records', []);
-  const onChange = useCallback(
-    applyFiltersAdapter(applyFilters),
-    [applyFilters],
-  );
 
   return (
     <OrderLinesFilters
       funds={funds}
       materialTypes={materialTypes}
       activeFilters={activeFilters}
-      onChange={onChange}
+      applyFilters={applyFilters}
       disabled={disabled}
     />
   );