+++ src/OrderLinesList/OrderLinesFiltersContainer.test.js
@@ -1,10 +1,12 @@
 import React from 'react';
 import { render, screen } from '@testing-library/react';
-import OrderLinesFilters from '@folio/plugin-find-po-line/FindPOLine/OrderLinesFilters';
+import { OrderLinesFilters } from '@folio/plugin-find-po-line';
 
 import OrderLinesFiltersContainer from './OrderLinesFiltersContainer';
 
-jest.mock('@folio/plugin-find-po-line/FindPOLine/OrderLinesFilters', () => jest.fn().mockReturnValue('OrderLinesFilters'));
+jest.mock('@folio/plugin-find-po-line/FindPOLine/OrderLinesFilters', () => ({
+  OrderLinesFilters: jest.fn().mockReturnValue('OrderLinesFilters'),
+}));
 
 const defaultProps = {
   activeFilters: {},
@@ -27,10 +29,10 @@ describe('OrderLinesFiltersContainer', () => {
     expect(screen.getByText('OrderLinesFilters')).toBeDefined();
   });
 
-  it('should call \'applyFilters\' function when filtes changed', () => {
+  it('should call \'applyFilters\' function when filters changed', () => {
     renderOrderLinesFiltersContainer();
 
-    OrderLinesFilters.mock.calls[0][0].onChange({ name: 'name', values: [] });
+    OrderLinesFilters.mock.calls[0][0].applyFilters({ name: 'name', values: [] });
 
     expect(defaultProps.applyFilters).toHaveBeenCalled();
   });