+++ src/OrderLinesList/hooks/index.js
@@ -1,3 +1,1 @@
-export * from './useLinesQuery';
-export * from './useNormalizedISBN';
 export * from './useOrderLines';